Add files via upload
Made these game over screens in the style of a BSOD as it  felt natural to me. One is a "clean" BSOD and the other has some dead pixel crash pattern lines added. I also threw the Exit button on both screens to give us an idea of what the game over screens would look like.